requirements.txt

kivy>=2.3.1
numpy>=1.21.0
tensorflow-lite>=2.10.0
Pillow>=9.0.0
opencv-python-headless>=4.5.0
